Added zoom, add/delete views, changed to drag and drop, and fixed bugs
Changed how panning works to allow for zooming. Implemented zooming based around the position of the mouse. Items are now moved by dragging them instead of picking them up. Changed the system for saving views and added the ability to rename, save, and delete views. Fixed many bugs, especially related to deleting items.

+++ src/options.py
@@ -2,47 +2,62 @@ import tkinter as tk
 from config import *
 
 class Options:
-    def __init__(self, root, block):
+    def __init__(self, root, to_configure):
         super().__init__()
         self.__root = root
-        self.__block = block
+        self.__to_configure = to_configure
         self.__window = tk.Toplevel(root)
         
         # self.__window.geometry(f"{WINDOW_WIDTH}x{WINDOW_HEIGHT}")
         self.__window.title("Options")
         self.__window.config(bg=BACKGROUND_COLOR)
         
-    def get_block(self):
-        return self.__block
+    def get_to_configure(self):
+        return self.__to_configure
         
     def get_window(self):
         return self.__window
         
     def add_entry(self, text):
         self.__entry_name_text = tk.StringVar()
-        self.__entry_name_text.set(self.__block.get_name())
+        self.__entry_name_text.set(self.__to_configure.get_name())
         
         self.add_label(0, 0, 1, text)
         
-        entry_name = tk.Entry(self.__window, textvariable=self.__entry_name_text)
+        entry_name = tk.Entry(self.__window, textvariable=self.__entry_name_text, font=FONT)
         entry_name.grid(row=0, column=1, padx=OPTION_FIELDS_PADDING, pady=OPTION_FIELDS_PADDING)
         
-        self.__entry_name_text.trace("w", self.update_block_name)
+        self.__entry_name_text.trace("w", self.update_name)
         
     def add_label(self, row, column, columnspan, text):
-        label = tk.Label(self.__window, text=text, bg=BACKGROUND_COLOR)
+        label = tk.Label(self.__window, text=text, font=FONT, bg=BACKGROUND_COLOR)
         label.grid(row=row, column=column, columnspan=columnspan, padx=OPTION_FIELDS_PADDING, pady=OPTION_FIELDS_PADDING)
         
     def add_delete_button(self, row, column, columnspan, to_delete):
-        delete_button = tk.Button(self.__window, text="Delete", command=lambda: self.delete(to_delete))
+        delete_button = tk.Button(self.__window, text="Delete", font=FONT, command=lambda: self.delete(to_delete))
         delete_button.grid(row=row, column=column, columnspan=columnspan, sticky=tk.W+tk.E)
         
-    def update_block_name(self, *args):
-        self.get_block().set_name(self.__entry_name_text.get())
+    def update_name(self, *args):
+        self.__to_configure.set_name(self.__entry_name_text.get())
         
     def delete(self, to_delete):
-        to_delete.delete()
+        if to_delete != None:
+            to_delete.delete()
+            
         self.__window.destroy()
+        
+class OptionsView(Options):
+    def __init__(self, model, view):
+        super().__init__(model.get_root(), view)
+        self.__model = model
+        self.__view = view
+        
+        self.add_entry("Name:")
+        self.add_delete_button(1, 0, 2, None)
+        
+    def delete(self, to_delete):
+        super().delete(to_delete)
+        self.__model.delete_view(self.__view)
         
 class OptionsConfigurationClass(Options):
     def __init__(self, root, configuration_class_gui):
@@ -60,13 +75,16 @@ class OptionsSetupClass(Options):
         
         self.add_entry("Name:")
         
-        self.add_label(0, 0, 1, "Add linked copy to setup view:")
+        self.add_label(1, 0, 2, "Add linked copy to setup view:")
         
         for i, setup_view_name in enumerate(setup_view_names):
-            link_button = tk.Button(self.get_window(), text=setup_view_name, command=lambda i=i: self.create_linked_setup_class_gui(i))
-            link_button.grid(row=1+(i//2), column=i%2, sticky=tk.W+tk.E)
+            link_button = tk.Button(self.get_window(), text=setup_view_name, font=FONT, command=lambda i=i: self.create_linked_setup_class_gui(i))
+            link_button.grid(row=2+i, columnspan=2, sticky=tk.W+tk.E)
             
-        self.add_delete_button(1+int(len(setup_view_names)+0.5), 0, 2, setup_class_gui)
+        row_after_link_buttons = 2 + len(setup_view_names)
+        
+        self.add_label(row_after_link_buttons, 0, 2, "")
+        self.add_delete_button(row_after_link_buttons+1, 0, 2, setup_class_gui)
         
     def create_linked_setup_class_gui(self, setup_view_number):
         self.__model.create_linked_setup_class_gui(self.__setup_class_gui, self.__configuration_class_gui, setup_view_number)
@@ -85,34 +103,22 @@ class OptionsConfigurationAttribute(Options):
         for i, value_type_symbol_and_text in enumerate(ACTIVE_VALUE_TYPE_SYMBOLS_CONFIGS):
             value_type_symbol, text = value_type_symbol_and_text
             
-            radio_button = tk.Radiobutton(self.get_window(), text=text, variable=self.__value_type, value=value_type_symbol, command=self.set_value_type, bg=BACKGROUND_COLOR, width=OPTION_RADIO_BUTTON_CONFIGURATION_ATTRIBUTE_WIDTH, anchor="w")
+            radio_button = tk.Radiobutton(self.get_window(), text=text, font=FONT, variable=self.__value_type, value=value_type_symbol, command=self.set_value_type, bg=BACKGROUND_COLOR, width=OPTION_RADIO_BUTTON_CONFIGURATION_ATTRIBUTE_WIDTH, anchor="w")
             radio_button.grid(row=2+i, columnspan=2, padx=OPTION_FIELDS_PADDING, pady=OPTION_FIELDS_PADDING, sticky=tk.W+tk.E)
             
         row_after_radio_buttons = 2 + len(ACTIVE_VALUE_TYPE_SYMBOLS_CONFIGS)
         self.add_label(row_after_radio_buttons, 0, 2, "")
         
-        self.__hidden_toggle = tk.Checkbutton(self.get_window(), text="Hide", variable=self.__is_hidden, command=self.set_is_hidden)
+        self.__hidden_toggle = tk.Checkbutton(self.get_window(), text="Hide", font=FONT, variable=self.__is_hidden, command=self.set_is_hidden)
         self.__hidden_toggle.grid(row=row_after_radio_buttons+1, columnspan=2, padx=OPTION_FIELDS_PADDING, pady=OPTION_FIELDS_PADDING, sticky=tk.W+tk.E)
         
         self.add_delete_button(row_after_radio_buttons+2, 0, 2, configuration_attribute_gui)
-        
-        """
-        calculation_inputs = attribute_block.get_calculation_inputs()
-        
-        self.__label_value = tk.Label(self.get_window(), text="Value:", bg=BACKGROUND_COLOR)
-        self.__label_value.grid(row=1, column=0, padx=OPTION_FIELDS_PADDING, pady=OPTION_FIELDS_PADDING)
-        
-        if len(calculation_inputs) == 0:
-            self.__entry_value = tk.Entry(self.get_window())
-            self.__entry_value.grid(row=1, column=1, padx=OPTION_FIELDS_PADDING, pady=OPTION_FIELDS_PADDING)
-            self.__entry_value.insert(0, "SOMETHING")
-        """
         
     def set_value_type(self):
         self.__configuration_attribute_gui.set_value_type(self.__value_type.get())
         
     def set_is_hidden(self):
-        self.__attribute_block.set_hidden(self.__is_hidden.get())
+        self.__configuration_attribute_gui.set_hidden(self.__is_hidden.get())
             
 class OptionsCalculationInput(Options):
     def __init__(self, root, configuration_input):
@@ -122,7 +128,7 @@ class OptionsCalculationInput(Options):
         for i, config in enumerate(ACTIVE_CALCULATION_TYPE_SYMBOLS_CONFIGS):
             calculation_symbol, text = config
             
-            radio_button = tk.Radiobutton(self.get_window(), text=text, variable=self.__calculation_type_choice, value=calculation_symbol, command=self.set_calculation_type, bg=BACKGROUND_COLOR, width=OPTION_RADIO_BUTTON_CONFIGURATION_INPUT_WIDTH, anchor="w")
+            radio_button = tk.Radiobutton(self.get_window(), text=text, font=FONT, variable=self.__calculation_type_choice, value=calculation_symbol, command=self.set_calculation_type, bg=BACKGROUND_COLOR, width=OPTION_RADIO_BUTTON_CONFIGURATION_INPUT_WIDTH, anchor="w")
             radio_button.grid(row=i, padx=OPTION_FIELDS_PADDING, pady=OPTION_FIELDS_PADDING, sticky=tk.W+tk.E)
             
         row_after_radio_buttons = len(ACTIVE_CALCULATION_TYPE_SYMBOLS_CONFIGS)
@@ -131,7 +137,7 @@ class OptionsCalculationInput(Options):
         self.add_delete_button(row_after_radio_buttons+1, 0, 1, configuration_input)
         
     def set_calculation_type(self):
-        self.get_block().set_symbol_calculation_type(self.__calculation_type_choice.get())
+        self.get_to_configure().set_symbol_calculation_type(self.__calculation_type_choice.get())
         
 class OptionsConnection(Options):
     def __init__(self, root, connection):
@@ -139,7 +145,7 @@ class OptionsConnection(Options):
         self.__connection = connection
         self.__is_external = tk.BooleanVar()
         
-        self.__external_toggle = tk.Checkbutton(self.get_window(), text="External connection", variable=self.__is_external, command=self.set_is_external)
+        self.__external_toggle = tk.Checkbutton(self.get_window(), text="External connection", font=FONT, variable=self.__is_external, command=self.set_is_external)
         self.__external_toggle.grid(row=0, padx=OPTION_FIELDS_PADDING, pady=OPTION_FIELDS_PADDING, sticky=tk.W+tk.E)
         
         self.add_delete_button(1, 0, 1, self.__connection)
